Record turns and prevent flying(?)
1) Changes to record when the mouse goes down the wrong arm of the T
maze, and displays the rate of correct first turns toward the target.
If this increases over time, this is evidence of learning.

2) There is a bug where rarely (about 1 per hour) the mouse collides
with a wall and then gets elevated above the wall.  This is bad as the
mouse is stuck up there.  I disabled overlapRecovery in the character
controller to see if this might help.  The defect is hard to replicate
so I will run some mice and see.